requirements.txt
streamlit==1.48.1
pandas==2.3.2
openpyxl==3.1.5
ollama==0.5.3
matplotlib==3.10.5
behavex==4.5.1